+++ R/as_draws.R
@@ -3,15 +3,17 @@
 #' Converts the output from [dynamite::dynamite()] call to a
 #' `draws_df` format of the \pkg{posterior} package, enabling the use
 #' of diagnostics and plotting methods of \pkg{posterior} and \pkg{bayesplot}
-#' packages. Note that this function returns all variables in a wide format,
+#' packages. Note that this function returns variables in a wide format,
 #' whereas [dynamite::as.data.frame()] uses the long format.
 #'
-#' You can use the arguments `responses` and `types` to extract only a subset
-#' of the model parameters (i.e., only certain types of parameters related to a
-#' certain response variable).
+#' You can use the arguments `parameters`, `responses` and `types` to extract
+#' only a subset of the model parameters (i.e., only certain types of parameters
+#'  related to a certain response variable).
 #'
 #' See potential values for the types argument in
-#' [dynamite::as.data.frame.dynamitefit()]
+#' [dynamite::as.data.frame.dynamitefit()] and
+#' [dynamite::get_parameter_names()] for potential values for `parameters`
+#' argument.
 #' @export
 #' @aliases as_draws_df
 #' @export as_draws_df
@@ -22,12 +24,14 @@
 #' @examples
 #' as_draws(gaussian_example_fit, types = c("sigma", "beta"))
 #'
-as_draws_df.dynamitefit <- function(x, responses = NULL, types = NULL, ...) {
+as_draws_df.dynamitefit <- function(x, parameters = NULL, responses = NULL,
+  types = NULL, ...) {
   # avoid NSE notes from R CMD check
   .chain <- .draw <- .iteration <- NULL
   category <- group <- parameter <- response <- time <- type <- value <- NULL
   d <- as.data.table.dynamitefit(
     x,
+    parameters = parameters,
     responses = responses,
     types = types,
     summary = FALSE,