+++ R/coef.R
@@ -4,13 +4,15 @@
 #'
 #' @export
 #' @param object \[`dynamitefit`]\cr The model fit object.
+
 #' @param type  \[`character(1)`]\cr Either `beta` (the default) for
 #'   time-invariant coefficients, `delta` for time-varying coefficients,
 #'   `nu` for random effects, `lambda` for factor loadings, or `psi` for
-#'   latent factor.
+#'   latent factor. Ignored if the argument `parameters` is supplied.
 #' @param include_alpha \[`logical(1)`]\cr If `TRUE` (default), extracts also
 #'   time-invariant intercept term alpha if time-invariant parameters beta are
 #'   extracted, and time-varying alpha if time-varying delta are extracted.
+#'   Ignored if the argument `parameters` is supplied.
 #'  @param summary \[`logical(1)`]\cr If `TRUE` (default), returns posterior
 #'   mean, standard deviation, and posterior quantiles (as defined by the
 #'   `probs` argument) for all parameters. If `FALSE`, returns the
@@ -25,28 +27,33 @@
 #' deltas <- coef(gaussian_example_fit, type = "delta")
 #'
 coef.dynamitefit <- function(object,
+                             parameters = NULL,
                              type = c("beta", "delta", "nu", "lambda", "psi"),
                              responses = NULL, summary = TRUE,
                              probs = c(0.05, 0.95),
                              include_alpha = TRUE, ...) {
-  type <- onlyif(is.character(type), tolower(type))
-  type <- try(
-    match.arg(type, c("beta", "delta", "nu", "lambda", "psi")),
-    silent = TRUE
-  )
-  stopifnot_(
-    !inherits(type, "try-error"),
-    "Argument {.arg type} must be either \"beta\", \"delta\", \"nu\",
-    \"lambda\", or \"psi\"."
-  )
-  stopifnot_(
-    checkmate::test_flag(x = include_alpha),
-    "Argument {.arg include_alpha} must be single {.cls logical} value."
-  )
-  if (include_alpha && type %in% c("beta", "delta")) {
+  if (is.null(parameters)) {
+    type <- onlyif(is.character(type), tolower(type))
+    type <- try(
+      match.arg(type, c("beta", "delta", "nu", "lambda", "psi")),
+      silent = TRUE
+    )
+    stopifnot_(
+      !inherits(type, "try-error"),
+      "Argument {.arg type} must be either \"beta\", \"delta\", \"nu\",
+      \"lambda\", or \"psi\"."
+    )
+    stopifnot_(
+      checkmate::test_flag(x = include_alpha),
+      "Argument {.arg include_alpha} must be single {.cls logical} value."
+    )
+  }
+
+  if (is.null(parameters) && include_alpha && type %in% c("beta", "delta")) {
     types <- c("alpha", type)
     out <- as.data.frame.dynamitefit(
       object,
+      parameters = parameters,
       types = types,
       responses = responses,
       summary = summary,
@@ -61,6 +68,7 @@ coef.dynamitefit <- function(object,
   } else {
     out <- as.data.frame.dynamitefit(
       object,
+      parameters = parameters,
       types = type,
       responses = responses,
       summary = summary,